refactor: 기존의 메인 하단 컴포넌트들 components의 tabs폴더로 이동 및 Section붙임
