+++ conponents/(tabs)/NearActivitySection.tsx
